requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
m3u8==0.8.0
pycryptodome==3.17
selenium
soupsieve==2.5
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
tqdm
